Clarify that FLOAT4E2M1 can be in int32_data (#6640)
### Description
Clarify in spec proto that FLOAT4E2M1 can be in int32_data, according to
test usage.
Updated text in the spec for int32_data for better readability and
accuracy.

### Motivation and Context

Previously the spec was incomplete according to the added tests.
FLOAT4E2M1 was not yet released so the change should not require a new
IR version.

---------

Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>