Check the input of ConstantOfShape to be non-negative (#6691)
### Description
I added a flag parameter in the `getShapeInput` function to check shape
dims non-negative in some cases.

### Motivation and Context
This PR is to solve the issue #6135. It proposes that the operator
`ConstantOfShape` doesn't check input validation and may produce
negative output.

I added a flag to the `getShapeInput` function. When it's true, the
function will check the shape returned and fail if any aim is negative.
This flag is enabled on `ConstantOfShape` operator.

(By the way, it's my first time contributing a PR to the open-source
community. Please let me know if there is anything wrong with it. Thank
you for your attention and review🙂

---------

Signed-off-by: jinjieliu <jinjie.liu@usc.edu>
Signed-off-by: JinjieLiu <jinjie.liu@usc.edu>