check_for_publish_preview_build_to_testpypi_weekly in CI should fail for PullRequests (#6727)
### Description
the pipeline job "check_for_publish_preview_build_to_testpypi_weekly"
should check if all condition for publishing to testpyi are met.

This must be done in a separate job, as a job with a github-environement
first checks whether the authorizations for using the environment have
been checked.

Currently, the “check for publish” preview job does not yet check which
branch is to be used.

This has the following undesired behavior


![grafik](https://github.com/user-attachments/assets/503a427f-9c62-43e7-b1f5-e23a7ea6b486)

Signed-off-by: Andreas Fehlner <fehlner@arcor.de>