Add multi-device execution support in ONNX (#6641)
Co-authored with @gramalingam 

### Description
Updates ONNX specification to include fields to describe multi-device
inference execution.

### Motivation and Context
The recent trend in increasingly larger models has spurred an interest
in distributed inference. A key performance bottleneck for inference for
these large models has been the memory limits of GPUs and other
accelerators as well as communication bandwidth. Thus, efficient
distributed inference typically requires parallelization of the
computation across multiple devices taking memory and bandwidth into
account.

Our goal is to extend ONNX so that it can serve as a representation of a
parallelized model. This is driven by the current state-of-the-art
techniques used for distributed inference (eg., see[ GSPMD: General and
Scalable Parallelization for ML Computation
Graphs](https://arxiv.org/pdf/2105.04663.pdf)). In particular, two
techniques of interest are: tensor parallelism and pipelining. In tensor
parallelism (also known as horizontal parallelism or operator
parallelism), the computation of a single operator (node) in the graph
is parallelized across multiple devices by sharding its inputs, In
pipeline parallelism, different subgraphs are assigned to different
devices.

---------

Signed-off-by: Kevin Chen <kevinch@nvidia.com>
Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>
Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>