replace quansight-labs/setup-python with actions/setup-python (#6837)
Fixes #6805 

### Description
<!-- - Describe your changes. -->

Updates to actions/setup-python 5.5.0 and replaces
quansight-labs/setup-python with actions/setup-python where it is used.
Also gets rid of free-threaded specific workflows that are no longer
necessary.

### Motivation and Context
<!-- - Why is this change required? What problem does it solve? -->
<!-- - If it fixes an open issue, please link to the issue here. -->

We archived quansight-labs/setup-python this morning so I'm trying to
help projects who adopted our fork to migrate off of it now that it's no
longer needed.

---------

Signed-off-by: Nathan Goldbaum <nathan.goldbaum@gmail.com>