Update release_linux_aarch64.yml (#6876)
This pull request includes changes to the
`.github/workflows/release_linux_aarch64.yml` file to streamline the
naming and usage of the manylinux environment variable. The most
important changes include the addition of a new environment variable and
updates to the build steps to use this variable.

Improvements to workflow configuration:

* Added a new environment variable `manylinux` to the `jobs` section.
* Updated the build step name from "Build manylinux2014_aarch64" to
"Build manylinux".
* Modified the `entrypoint.sh` script execution to use the new
`manylinux` environment variable instead of hardcoding
`manylinux2014_aarch64`.

Signed-off-by: Andreas Fehlner <fehlner@arcor.de>