Use CONFIG to find protobuf (#6840)
### Description
This uses the CMake configuration file mechanism to configure protobuf. 

### Motivation and Context
This is needed to correctly account for the dependency on abseil of the
recent versions of protobuf.
Moreover this has the side effect that the generated file is actually
relocatable fine if we happen to build in one folder and deploy in
another.

Signed-off-by: Giulio Eulisse <10544+ktf@users.noreply.github.com>