Add support for Free-Threading in Linux Wheel Builds (#6835)
### Description
This pull request introduces support for building Linux wheels with
free-threading enabled. The changes involve modifying the build
configuration and ensuring compatibility with the ONNX framework.

### Motivation and Context
Adding free-threading support to Linux wheels improves performance and
scalability for multi-threaded applications that rely on ONNX. This
change addresses the need for better multi-threading capabilities in the
ONNX ecosystem and enhances its usability in high-performance computing
environments.

---------

Signed-off-by: Andreas Fehlner <fehlner@arcor.de>