Use Native Runner für linux Aarch64 whl (public preview runner available) (#6907)
### Description
This pull request introduces the use of a native runner for Linux
AArch64 wheel builds. By leveraging the public preview runner, it aims
to optimize the build process and improve overall efficiency.

### Motivation and Context
The current Linux AArch64 Docker-based build is slow and acts as a
bottleneck in the CI process. This change addresses the issue by
transitioning to a native runner, which is expected to significantly
speed up the build process and unblock CI pipelines.

---------

Signed-off-by: Andreas Fehlner <fehlner@arcor.de>
Signed-off-by: gramalingam <grama@microsoft.com>