fix broken links to automatic upgrade and downgrade tests in AddNewOp.md (#7023)
### Description
Minor fix so that links to the `automatic_upgrade_test.py` and
`automatic_downgrade_test.py` file links point to the
`test/version_converter` directory instead of just the `test/`
directory.

### Motivation and Context
Link was broken before which was confusing for understanding which files
need to be modified. This means readers of the document can now be
directed to the correct location.

Signed-off-by: Alexander Poulson <alexanderipoulson@gmail.com>