Correct `dtype` attribute docs for EyeLike operator (#7031)
The docs for EyeLike's `dtype` attribute read:

> The data type for the elements of the output tensor. If not
specified,the data type of the input tensor T1 is used. If input tensor
T1 is also notspecified, then type defaults to ‘float’.

The input tensor is not optional, so it doesn't make sense to specify a
fallback for the `dtype` attribute if it isn't provided. Remove this
last sentence.

Signed-off-by: Robert Knight <robertknight@gmail.com>
Co-authored-by: Andreas Fehlner <fehlner@arcor.de>