Add Pixi development environment (#7044)
### Description

This PR enables an easy, reproducible, and cross-platform (this PR ships
linux-64 and osx-arm64 support, but Windows can be added, too)
conda-forge-based developer environment using the
[Pixi](https://pixi.sh/latest/) package manager. A CI workflow testing
that onnx can be correctly built on MacOS arm64 and Linux x86-64 using
recent and minimal dependencies is also included.

### Motivation and Context

#### Improve developer experience

Sporadically contributing to onnx has been a pain point for me and my
colleagues in the past. The issues are primarily centered on the
following:
- No easily reproducible developer environment that is known to work
(based on lock files). This means compatible (non-Python) dependencies
such as protobuf, cmake, various linters, or compiler toolchains must be
manually maintained and updated.
- Required CMake flags and build commands are easy to forget and not
easy to discover. Examples in this regard are running the gtest suite or
re-rendering the operators' docs for the ai.onnx.ml domain.

Conda-forge is an established ecosystem, especially in data science and
academic communities. Pixi is a cross-platform, state-of-the-art package
manager based on Conda-forge. It allows for creating reproducible
developer environments based on a lock file and enables simple
orchestration tasks based on "tasks". See `INSTALL.md` edits for further
details.

#### CI testing

We have run into issues (a recent
[example](https://github.com/conda-forge/onnx-feedstock/pull/131))
building onnx for conda-forge in the past. While this PR does not
guarantee that such issues won't occur in the future, it would certainly
help to ensure in the CI of this project that things still work on the
conda-forge infrastructure. Pixi is very well suited for CI tasks. It
allows for the efficient definition of environments with different
dependencies which can then be used in CI. For instance, this PR ensures
that onnx can be correctly built in an environment with (locked) old
dependencies ("oldies" environment) and (locked) recent dependencies
("default" environment).

While not a replacement for ManyLinux-based PyPI builds, the CI
integration in this PR may serve as a starting point to further simplify
CI workflows.

---------

Signed-off-by: Christian Bourjau <christian.bourjau@quantco.com>
Co-authored-by: Andreas Fehlner <fehlner@arcor.de>