Update input and output tensors in pb files to match the model (#7074)
There are a few tests that are consistently
[failing](https://aiinfra.visualstudio.com/PublicPackages/_build/results?buildId=831429&view=logs&j=9d976e38-31ec-50dd-b1f8-279fbf889fca&t=85ed5ad3-b72a-52c3-abe0-a87b66004fd0)
on the QNN CI Pipelines triggered for
https://github.com/microsoft/onnxruntime/pull/24973.

For each of these tests, we see this error in the logs: 
`[E:onnxruntime:Default, dataitem_request.cc:32 operator()]
<test_name>:Invalid input name: x`.

This seems to happen due to a mismatch between the name of the tensors
in the `model.onnx` file of each test (which are all `input` and
`output`) and the name of the tensors in the .pb files. This PR updates
the `input_0.pb` and `output_0.pb` files for each failing test to rename
the input tensor from `x` to `input` and output tensor from `y` to
`output`.

Signed-off-by: Alex Marin <emarin@microsoft.com>