Remove re2 dependency (#7083)
### Description
My understanding is that re2 offers a smaller feature set than Python's
re module. The ONNX specs for `RegexFullMatch` call for RE2 syntax.
Thus, replacing re2 with Python's own re package actually broadens the
support beyond what the specs call for. While not idea, I believe that
trade-off is reasonable given the unresponsiveness of the google-re2
project in resolving current issues around Python 3.13.

### Motivation and Context

`google-re2` (i.e. the Python bindings to the re2 library) appear
somewhat unmaintained. There is a long-open
[issue](https://github.com/google/re2/issues/516) about support for
Python 3.13 which is being ignored. Furthermore, the package does not
officially support Windows and is not available on conda-forge on that
platform.

---------

Signed-off-by: Christian Bourjau <christian.bourjau@quantco.com>
Co-authored-by: Andreas Fehlner <fehlner@arcor.de>