Add TensorScatter op for in-place kv cache update (#7114)
### Description
This PR adds TensorScatter op for in-place kv updates, to be used in
Attention computations.

This op takes the same shape for the input past_cache and output
present_cache (with the sequence length dimension being max_seqlen) so
that backends are free to alias them to achieve efficient in-place
updates during the autoregressive iterations of Transformer models.

### Motivation and Context
This work is part of the effort of the GenAI WG to enable LLM features
in ONNX.

---------

Signed-off-by: Yuan Yao <yuanyao@nvidia.com>
Signed-off-by: Yash solanki <alphacr792@gmail.com>