Add kv_nonpad_seqlen input to Attention (#7164)
### Description

To accompany the
[TensorScatter-24](https://github.com/onnx/onnx/pull/7114) op for
managing in-place KV cache update, this PR makes the following changes
to the Attention op:
- Add `nonpad_kv_seqlen` to indicate the number of valid (nonpadded)
tokens in the K and V inputs when the K and V inputs are the entire
cache tensors (where the number of valid tokens can potentially make up
only a small proportion of the cache tensors). The `nonpad_kv_seqlen`
input would provided optimization opportunities for backends to skip the
unnecessary computation on the padding tokens.
- Allow the kv_seqlen dimension (-1 dimension) of `attn_mask` input to
be shorter than K and V. The missing portion will be assumed to be -inf.
The length should still be larger than the max value in
`nonpad_kv_seqlen`.

Also, allow `attn_mask` and `is_causal` to be present at the same time.
This would allow for easier export of HF models later.

### Motivation and Context
<!-- - Why is this change required? What problem does it solve? -->
<!-- - If it fixes an open issue, please link to the issue here. -->

---------

Signed-off-by: Yuan Yao <yuanyao@nvidia.com>