Fix division for integer dtype in reference implementation (#7203)
### Description
So far, the division operator in the reference implementation worked via
``np.divide``, which produces a float dtype when both inputs have an
integer input dtype. the result was then cast back to integer dtype.
Now, integer division is put in place when both inputs are integers.

### Motivation and Context
We came over this bug when testing the reference implementation vs. the
onnxruntime in https://github.com/cbourjau/onnx-tests.

---------

Signed-off-by: Konstantin Pueckler <konstantin.pueckler@quantco.com>
Signed-off-by: konstantin-pueckler-qc <konstantin.pueckler@quantco.com>