Update pixi.{toml,lock} and create issues if CI fails (#7280)
### Description
This PR updates the `pixi.{toml,lock}` files to reflect the latest
dependencies of this project. It furthermore updates the weekly CI
action such that it automatically creates an issue and assigns me to it
on failure.

### Motivation and Context
The pixi-based developer environment was outdated for a couple of weeks
without me noticing. Thanks for the nuddge, @justinchuby

Signed-off-by: Christian Bourjau <christian.bourjau@quantco.com>