Adjust remaining tests which are using tobytes().
On big endian machines, we have to ensure that the bytes stored in a
tensor are converted to little endian.  While searching in the sources,
those are the remaining places - all testcases - where tobytes() is
used.  This patch either add a byteswap or is now passing a numpy array
to make_tensor, which then performs the byteswap, instead of using
tobytes().

Note that all corresponding tests did not fail before.  Thus in theory
we do not need those changes.  But it is more consistent as before.

Signed-off-by: Stefan Liebler <stli@linux.ibm.com>