Fix testsuite fails on s390x (big-endian) (#7309)
On big-endian machines, there are plenty of test fails due to
missing/wrong byteswaps. I've created a couple of patches to fix those
fails, in numpy_helper.py::from_array() and helper.py::make_tensor() and
in various testcases.

Some of the fixed Testcases test_make_tensor_raw are also mentioned in
the issue from 2024-07-14 created by tehbone:
[s390x test failures - test_make_tensor_raw
#6181](https://github.com/onnx/onnx/issues/6181)

Especially the last commit "Adjust remaining tests which are using
tobytes()." is not really required as all corresponding tests did not
fail before. But it makes the tests more consistent as before.

---------

Signed-off-by: Stefan Liebler <stli@linux.ibm.com>
Signed-off-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>
Co-authored-by: justinchuby <11205048+justinchuby@users.noreply.github.com>