Fix node level inference for GreaterOrEqual op (#7428)
Fix a couple of issues.

* Node-level inference invoked from python was not working for
GreaterOrEqual. Looks like the type-and-shape-inference method was not
set for GreaterOrEqual and LessOrEqual operators. Add this.
* Even without above, the inference should have inferred the output type
(from the schema). This was not working because of a test on the python
side that skipped calling the C++ inference logic if no
type-and-shape-inference method was set for the op. But the inference
logic has type-inference logic (from schema) independent of the
shape-inference method which should be called anyway.

One more potential fix for the future: even without an explicit
shape-inference method, the function-expansion could have been used for
these function-ops. The full shape-inference logic does this, I believe,
but this is not exposed in the node-level inference method ... but it
could be integrated here too. Leaving this for a future PR.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>